+++ render/template/engine/html/html.go
@@ -1,4 +1,4 @@
-package standar
+package html
 
 import (
 	"bytes"
@@ -21,28 +21,6 @@ var (
 )
 
 type (
-	/*
-	 this was the old good way before 3.0.0-alpha.1, but after I released that this brings troubles to simplify the Configuration for the iris' station
-	 so I moved all configs inside the config package and all are happy,
-	 the only downside here is that, we will have the same config struct for all template engines
-	 and each template engine will pick its config to work on,
-	 standar will pick the config.Standar, pongo the config.Pongo, and the common options are in config.
-	*/
-	/*
-		StandarConfig struct {
-			RequirePartials bool
-			// Delims
-			Left  string
-			Right string
-			// Funcs for Standar
-			Funcs []template.FuncMap
-		}
-
-		Config struct {
-			engine.Config
-			StandarConfig
-		}
-	*/
 	Engine struct {
 		Config    *config.Template
 		Templates *template.Template
@@ -63,13 +41,7 @@ var emptyFuncs = template.FuncMap{
 	},
 }
 
-/*
-func WrapConfig(common engine.Config, standar StandarConfig) *Config {
-	return &Config{Config: common, StandarConfig: standar}
-}
-*/
-
-// New creates and returns a Standar template engine
+// New creates and returns a HTMLTemplate  engine
 func New(cfg ...config.Template) *Engine {
 	if buffer == nil {
 		buffer = utils.NewBufferPool(64)
@@ -102,7 +74,7 @@ func (s *Engine) buildFromDir() error {
 	var templateErr error
 	dir := s.Config.Directory
 	s.Templates = template.New(dir)
-	s.Templates.Delims(s.Config.Standar.Left, s.Config.Standar.Right)
+	s.Templates.Delims(s.Config.HTMLTemplate.Left, s.Config.HTMLTemplate.Right)
 
 	// Walk the supplied directory and compile any files that match our extension list.
 	filepath.Walk(dir, func(path string, info os.FileInfo, err error) error {
@@ -135,7 +107,7 @@ func (s *Engine) buildFromDir() error {
 				tmpl := s.Templates.New(name)
 
 				// Add our funcmaps.
-				for _, funcs := range s.Config.Standar.Funcs {
+				for _, funcs := range s.Config.HTMLTemplate.Funcs {
 					tmpl.Funcs(funcs)
 				}
 
@@ -153,7 +125,7 @@ func (s *Engine) buildFromAsset() error {
 	var templateErr error
 	dir := s.Config.Directory
 	s.Templates = template.New(dir)
-	s.Templates.Delims(s.Config.Standar.Left, s.Config.Standar.Right)
+	s.Templates.Delims(s.Config.HTMLTemplate.Left, s.Config.HTMLTemplate.Right)
 
 	for _, path := range s.Config.AssetNames() {
 		if !strings.HasPrefix(path, dir) {
@@ -182,7 +154,7 @@ func (s *Engine) buildFromAsset() error {
 				tmpl := s.Templates.New(name)
 
 				// Add our funcmaps.
-				for _, funcs := range s.Config.Standar.Funcs {
+				for _, funcs := range s.Config.HTMLTemplate.Funcs {
 					tmpl.Funcs(funcs)
 				}
 
@@ -212,7 +184,7 @@ func (s *Engine) layoutFuncsFor(name string, binding interface{}) {
 		},
 		"partial": func(partialName string) (template.HTML, error) {
 			fullPartialName := fmt.Sprintf("%s-%s", partialName, name)
-			if s.Config.Standar.RequirePartials || s.Templates.Lookup(fullPartialName) != nil {
+			if s.Config.HTMLTemplate.RequirePartials || s.Templates.Lookup(fullPartialName) != nil {
 				buf, err := s.executeTemplateBuf(fullPartialName, binding)
 				// Return safe HTML here since we are rendering our own template.
 				return template.HTML(buf.String()), err